(registration-react) refactor containers to pages
