+++ applications/registration-react/src/pages/catalogs-page/catalogs-page.scss
