+++ applications/registration-react/src/pages/dataset-list-page/dataset-list-page.scss
