+++ applications/registration-react/src/pages/dataset-registration-page/dataset-registration-page.scss
